IE Fix. Stringify/Parse delta sent across bridge.
IE's execute_script does not handle js objects well; it randomly overwrites
properties with null. Stringifying the object in JS and parsing it in Ruby gets
around this problem, though with a bit of a performance hit. If it's too slow,
I'll rewrite the code to only do this in IE.